% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_bind_xlsx.R
\name{read_and_bind_xlsx}
\alias{read_and_bind_xlsx}
\title{read_and_bind_xlsx (package "deptstat")}
\usage{
read_and_bind_xlsx(path, filename_pattern = "", return_filenames = F)
}
\arguments{
\item{path}{A path to XLSX files}

\item{filename_pattern}{A substring for subsetting files (part of file name)}

\item{return_filenames}{A parameter, that can be set to TRUE if the output must contain source filenames. Default: FALSE}
}
\value{
A vector or data frame with outliers marked as 1 and normal observations marked as 0
}
\description{
This function is created to bind similar files into one data frame.
}
\examples{
# This will return a dataframe, based on binded XLSX files, which name contained 'data' as a substring
read_and_bind_xlsx(getwd(), "data")

#'  @author Dmitrii Diachkov (2021)

}
